+++ firmware/cartridges/c64gs_system_3.c
@@ -21,7 +21,7 @@
 /*************************************************
 * C64 bus read callback
 *************************************************/
-static inline bool c64gs_read_handler(uint8_t control, uint16_t addr)
+static inline bool c64gs_read_handler(uint32_t control, uint32_t addr)
 {
     if (!(control & C64_ROML))
     {
@@ -41,7 +41,7 @@ static inline bool c64gs_read_handler(uint8_t control, uint16_t addr)
 /*************************************************
 * C64 bus write callback
 *************************************************/
-static inline void c64gs_write_handler(uint8_t control, uint16_t addr, uint8_t data)
+static inline void c64gs_write_handler(uint32_t control, uint32_t addr, uint32_t data)
 {
     if (!(control & C64_IO1))
     {
