+++ firmware/cartridges/final_cartridge_3.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 Kim Jørgensen
+ * Copyright (c) 2019-2020 Kim Jørgensen
  *
  * This software is provided 'as-is', without any express or implied
  * warranty.  In no event will the authors be held liable for any damages
@@ -17,6 +17,7 @@
  *    misrepresented as being the original software.
  * 3. This notice may not be removed or altered from any source distribution.
  */
+
 static uint32_t const fc3_mode[8] =
 {
     STATUS_LED_ON|CRT_PORT_16K,
@@ -32,12 +33,12 @@ static uint32_t const fc3_mode[8] =
 
 #define FC3_REGISTER_ADDR (0xdfff)
 #define FC3_REGISTER_HIDE (0)
-static uint16_t fc3_register;
+static uint32_t fc3_register;
 
 /*************************************************
 * C64 bus read callback (VIC-II cycle)
 *************************************************/
-static inline bool fc3_vic_read_handler(uint8_t control, uint16_t addr)
+static inline bool fc3_vic_read_handler(uint32_t control, uint32_t addr)
 {
     if ((control & (C64_IO1|C64_IO2|C64_ROML|C64_ROMH)) != (C64_IO1|C64_IO2|C64_ROML|C64_ROMH))
     {
@@ -51,7 +52,7 @@ static inline bool fc3_vic_read_handler(uint8_t control, uint16_t addr)
 /*************************************************
 * C64 bus read callback (CPU cycle)
 *************************************************/
-static inline bool fc3_read_handler(uint8_t control, uint16_t addr)
+static inline bool fc3_read_handler(uint32_t control, uint32_t addr)
 {
     if ((control & (C64_IO1|C64_IO2|C64_ROML|C64_ROMH)) != (C64_IO1|C64_IO2|C64_ROML|C64_ROMH))
     {
@@ -89,7 +90,7 @@ static inline void fc3_early_write_handler(void)
 /*************************************************
 * C64 bus write callback
 *************************************************/
-static inline void fc3_write_handler(uint8_t control, uint16_t addr, uint8_t data)
+static inline void fc3_write_handler(uint32_t control, uint32_t addr, uint32_t data)
 {
     /*  The FC3 register is reset to $00 on reset.
         Bit 7: Hide register, 1 = disable write to register
