+++ firmware/cartridges/zaxxon.c
@@ -22,7 +22,7 @@
 /*************************************************
 * C64 bus read callback
 *************************************************/
-static inline bool zaxxon_read_handler(uint8_t control, uint16_t addr)
+static inline bool zaxxon_read_handler(uint32_t control, uint32_t addr)
 {
     if (!(control & C64_ROML))
     {
@@ -46,7 +46,7 @@ static inline bool zaxxon_read_handler(uint8_t control, uint16_t addr)
 /*************************************************
 * C64 bus write callback
 *************************************************/
-static inline void zaxxon_write_handler(uint8_t control, uint16_t addr, uint8_t data)
+static inline void zaxxon_write_handler(uint32_t control, uint32_t addr, uint32_t data)
 {
     // No write support
 }
